+++ src/ecc/ECC.jl
@@ -49,7 +49,7 @@ export parity_checks, parity_matrix_x, parity_matrix_z, iscss,
     evaluate_decoder,
     CommutationCheckECCSetup, NaiveSyndromeECCSetup, ShorSyndromeECCSetup,
     TableDecoder,
-    BeliefPropDecoder, BitFlipDecoder,
+    BeliefPropDecoder, BitFlipDecoder, BPOTSDecoder,
     PyBeliefPropDecoder, PyBeliefPropOSDecoder, PyMatchingDecoder, DecoderCorrectionGate
 
 """Parity check tableau of a code.